Custom waypoint movement example
This test sample allows the user to enter any number of waypoints with their corresponding latitude, longitude and altitude for autonomous movement of the drone. The drone will move to each waypoint in order as entered in and then return to the launch coordinates once done. This test is an easy way to move the drone to the exact coordinates you want it to visit.