Finalized custom example
This flight test example allows the user to enter in any number of waypoints with their respective latitude, longitude and altitude. Once points are entered the drone will move to each waypoint and stop for a second in between each point to verify it reached its destination. After the drone has reached the last waypoint, it will return to the launch site.